+++ core/src/main/java/io/miragon/bpmrepo/core/artifact/api/transport/ArtifactUpdateTO.java
@@ -24,4 +24,7 @@ public class ArtifactUpdateTO {
     @Nullable
     private String fileType;
 
+    @Nullable
+    private String file;
+
 }