+++ core/src/main/java/io/miragon/bpmrepo/core/artifact/domain/model/ArtifactUpdate.java
@@ -12,5 +12,6 @@ public class ArtifactUpdate {
     private String name;
     private String description;
     private String fileType;
+    private String file;
 
 }