+++ core/src/test/java/io/miragon/bpmrepo/core/DefaultFileTypesPlugin.java
@@ -23,16 +23,16 @@ public class DefaultFileTypesPlugin implements ArtifactTypesPlugin {
     public List<ArtifactTypeTO> getArtifactTypes() {
         final List<ArtifactTypeTO> fileTypes = new ArrayList<>();
 
-        ArtifactTypeTO type = new ArtifactTypeTO("BPMN", "bpmn", "settings", "modeler");
+        ArtifactTypeTO type = new ArtifactTypeTO("BPMN", "bpmn", "settings", "modeler", true, true);
         fileTypes.add(type);
 
-        type = new ArtifactTypeTO("DMN", "dmn", "view_list", "modeler");
+        type = new ArtifactTypeTO("DMN", "dmn", "view_list", "modeler",true, true);
         fileTypes.add(type);
 
-        type = new ArtifactTypeTO("FORM", "json", "reorder", "formulare");
+        type = new ArtifactTypeTO("FORM", "json", "reorder", "formulare",true, true);
         fileTypes.add(type);
 
-        type = new ArtifactTypeTO("CONFIGURATION", "json", "code", "konfiguration");
+        type = new ArtifactTypeTO("CONFIGURATION", "json", "code", null, false, true);
         fileTypes.add(type);
 
         return fileTypes;