+++ core/src/test/java/io/miragon/bpmrepo/core/artifact/DeploymentServiceTest.java
@@ -35,9 +35,10 @@ class DeploymentServiceTest {
     private static final String ARTIFACTDESC = "SomeDescription";
     private static final LocalDateTime DATE = LocalDateTime.now();
     private static final String USERNAME = "USER";
+    private static final String FILE_TYPE = "BPMN";
 
     public Artifact createArtifact() {
-        final Artifact artifact = ArtifactBuilder.buildArtifact(ARTIFACTID, REPOID, ARTIFACTNAME, ARTIFACTDESC, DATE, DATE);
+        final Artifact artifact = ArtifactBuilder.buildArtifact(ARTIFACTID, REPOID, ARTIFACTNAME, ARTIFACTDESC, FILE_TYPE, DATE, DATE);
         final Artifact createdArtifact = this.artifactService.createArtifact(artifact);
         assertNotNull(createdArtifact);
         return createdArtifact;