+++ example/src/main/java/io/miragon/bpmrepo/example/DefaultUserContext.java
@@ -8,6 +8,6 @@ public class DefaultUserContext implements UserContext {
 
     @Override
     public String getUserName() {
-        return "testuser";
+        return "demo.user";
     }
 }