+++ spring-boot-starter/src/main/java/io/miragon/bpmrepo/spring/boot/starter/fileTypes/DefaultFileTypesPlugin.java
@@ -13,16 +13,19 @@ public class DefaultFileTypesPlugin implements ArtifactTypesPlugin {
     public List<ArtifactTypeTO> getArtifactTypes() {
         final List<ArtifactTypeTO> fileTypes = new ArrayList<>();
 
-        ArtifactTypeTO type = new ArtifactTypeTO("BPMN", "bpmn", "settings", "modeler");
+        ArtifactTypeTO type = new ArtifactTypeTO("BPMN", "bpmn", "settings", "modeler", true, true);
         fileTypes.add(type);
 
-        type = new ArtifactTypeTO("DMN", "dmn", "view_list", "modeler");
+        type = new ArtifactTypeTO("DMN", "dmn", "view_list", "modeler", true, true);
         fileTypes.add(type);
 
-        type = new ArtifactTypeTO("FORM", "json", "reorder", "formulare");
+        type = new ArtifactTypeTO("FORM", "json", "reorder", "formulare", true, true);
         fileTypes.add(type);
 
-        type = new ArtifactTypeTO("CONFIGURATION", "json", "code", "editor");
+        type = new ArtifactTypeTO("ELEMENT_TEMPLATE", "json", "content_copy", null, false, false);
+        fileTypes.add(type);
+
+        type = new ArtifactTypeTO("CONFIGURATION", "json", "code", null, false, true);
         fileTypes.add(type);
 
         return fileTypes;